refactor: modularize mobile menu
